valueObjectに移動
