+++ src/domain/valueObject/label.go
@@ -1,0 +1,9 @@
+package valueObject
+
+type Label struct {
+	Id          uint16
+	Url         string
+	Name        string
+	Description string
+	Color       string
+}
